refactor: move ROI export helpers to PolygonImportExport
