+++ src/stores/PolygonDrawingStore/PolygonDrawingStore.helpers.ts
@@ -1,15 +1,4 @@
-import { useTranscriptLayerStore } from '../TranscriptLayerStore';
-import { useCellSegmentationLayerStore } from '../CellSegmentationLayerStore/CellSegmentationLayerStore';
-import { useViewerStore } from '../ViewerStore';
 import { PolygonFeature, Point2D, LineSegment, IntersectionResult, BoundingBox } from './PolygonDrawingStore.types';
-import {
-  CellsExportData,
-  TranscriptsExportData,
-  CellExportRecord,
-  RoiCellExport,
-  RoiTranscriptExport
-} from '../../components/PolygonImportExport/PolygonImportExport.types';
-import { SingleMask } from '../../shared/types';
 
 // Epsilon for floating point comparisons
 const EPSILON = 1e-10;
@@ -257,166 +246,6 @@ export const updatePolygonFeaturesWithIds = (features: PolygonFeature[], _nextId
     featuresWithIds,
     nextPolygonId: currentMaxId + 1
   };
-};
-
-const generateExportJsonFilename = (type: string): string => {
-  const viewerSource = useViewerStore.getState().source;
-  const ometiffName = viewerSource?.description?.replace(/\.(ome\.tiff?|tiff?|zarr)$/i, '') || 'export';
-  const date = new Date().toISOString().split('T')[0];
-  return `${ometiffName}_${type}_${date}.json`;
-};
-
-// Builds the canonical per-ROI export data consumed by both the JSON (below) and CSV serializers.
-export const buildTranscriptExports = (
-  polygonFeatures: PolygonFeature[],
-  polygonNotes: Record<number, string> = {}
-): RoiTranscriptExport[] => {
-  const { selectedPoints } = useTranscriptLayerStore.getState();
-
-  return polygonFeatures.map((feature) => {
-    const polygonId = feature.properties?.polygonId || 1;
-    const data = selectedPoints.find((selection) => selection.roiId === polygonId)?.data || [];
-
-    return {
-      roiName: `ROI_${polygonId}`,
-      polygonId,
-      coordinates: feature.geometry.coordinates[0].map((coord: number[]) => coord as [number, number]),
-      notes: polygonNotes[polygonId] || '',
-      transcripts: data.map((t) => ({
-        geneName: t.geneName || 'unknown',
-        position: t.position || [],
-        cellId: t.cellId || ''
-      }))
-    };
-  });
-};
-
-export const buildCellExports = (
-  polygonFeatures: PolygonFeature[],
-  includeGenes: boolean,
-  polygonNotes: Record<number, string> = {}
-): RoiCellExport[] => {
-  const { selectedCells, segmentationMetadata } = useCellSegmentationLayerStore.getState();
-
-  const proteinNames = segmentationMetadata?.proteinNames || [];
-  const allGeneNames = segmentationMetadata?.geneNames || [];
-  // An empty array is truthy, so gate gene data on length - keeps JSON and CSV in sync
-  const includeGeneData = includeGenes && allGeneNames.length > 0;
-  const geneNames = includeGeneData ? allGeneNames : [];
-
-  return polygonFeatures.map((feature) => {
-    const polygonId = feature.properties?.polygonId || 1;
-    const data = selectedCells.find((selection) => selection.roiId === polygonId)?.data || [];
-
-    return {
-      roiName: `ROI_${polygonId}`,
-      polygonId,
-      coordinates: feature.geometry.coordinates[0].map((coord: number[]) => coord as [number, number]),
-      notes: polygonNotes[polygonId] || '',
-      hasSegmentationMetadata: !!segmentationMetadata,
-      proteinNames,
-      geneNames,
-      cells: data.map((cell) => buildCellRecord(cell, proteinNames, geneNames, includeGeneData))
-    };
-  });
-};
-
-const buildCellRecord = (
-  cell: SingleMask,
-  proteinNames: string[],
-  geneNames: string[],
-  includeGeneData: boolean
-): CellExportRecord => {
-  const protein: Record<string, number> = {};
-  proteinNames.forEach((name, index) => {
-    protein[name] = cell.proteinValues[index];
-  });
-
-  let transcript: Record<string, number> | null = null;
-  if (includeGeneData) {
-    const transcriptMap: Record<string, number> = {};
-    cell.nonzeroGeneIndices.forEach((geneIndex, index) => {
-      const geneName = geneNames[geneIndex];
-      if (geneName !== undefined) {
-        transcriptMap[geneName] = cell.nonzeroGeneValues[index];
-      }
-    });
-    transcript = transcriptMap;
-  }
-
-  return {
-    cellId: cell.cellId,
-    area: cell.area,
-    totalCounts: cell.totalCounts,
-    totalGenes: cell.totalGenes,
-    clusterId: cell.clusterId,
-    vertices: cell.vertices,
-    umapValues: cell.umapValues,
-    protein,
-    transcript
-  };
-};
-
-export const exportPolygonsWithCells = (
-  polygonFeatures: PolygonFeature[],
-  includeGenes: boolean,
-  polygonNotes: Record<number, string>
-) => {
-  const exportData: CellsExportData = {};
-
-  buildCellExports(polygonFeatures, includeGenes, polygonNotes).forEach((roi) => {
-    exportData[roi.roiName] = {
-      coordinates: roi.coordinates,
-      cells: roi.cells.map((cell) => {
-        const { protein, transcript, ...rest } = cell;
-        return {
-          ...rest,
-          ...(roi.hasSegmentationMetadata ? { protein } : {}),
-          ...(transcript !== null ? { transcript } : {})
-        };
-      }),
-      polygonId: roi.polygonId,
-      notes: roi.notes
-    };
-  });
-
-  const blob = new Blob([JSON.stringify(exportData, null, 2)], { type: 'application/json' });
-  const url = URL.createObjectURL(blob);
-  const link = Object.assign(document.createElement('a'), {
-    href: url,
-    download: generateExportJsonFilename('segmentation')
-  });
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-  URL.revokeObjectURL(url);
-};
-
-export const exportPolygonsWithTranscripts = (
-  polygonFeatures: PolygonFeature[],
-  polygonNotes: Record<number, string>
-) => {
-  const exportData: TranscriptsExportData = {};
-
-  buildTranscriptExports(polygonFeatures, polygonNotes).forEach((roi) => {
-    exportData[roi.roiName] = {
-      coordinates: roi.coordinates,
-      transcripts: roi.transcripts,
-      polygonId: roi.polygonId,
-      notes: roi.notes
-    };
-  });
-
-  const blob = new Blob([JSON.stringify(exportData, null, 2)], { type: 'application/json' });
-  const url = URL.createObjectURL(blob);
-  const link = Object.assign(document.createElement('a'), {
-    href: url,
-    download: generateExportJsonFilename('transcripts')
-  });
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-  URL.revokeObjectURL(url);
 };
 
 export const findEditedPolygon = (
