+++ src/stores/PolygonDrawingStore/PolygonDrawingStore.types.ts
@@ -57,8 +57,6 @@ export type PolygonDrawingStoreMethods = {
   deletePolygon: (polygonId: number) => void;
   selectFeature: (index: number | null) => void;
   clearPolygons: () => void;
-  exportPolygonsWithCells: (includeGenes: boolean) => void;
-  exportPolygonsWithTranscripts: () => void;
   setDetecting: (detecting: boolean) => void;
   setPolygonOpacity: (opacity: number) => void;
   setPolygonNote: (polygonId: number, note: string) => void;
